refactor:Refactor: Adjust TaskCacheService to use interfaces and move implementations to infrastructure
