+++ src/domain/services/ITaskCacheService.ts
@@ -1,0 +1,4 @@
+export interface ITaskCacheService {
+  getTaskFromCache(taskId: string): Promise<any | null>;
+  setTaskInCache(taskId: string, taskData: any): Promise<void>;
+}
